+++ cp/migrations/000006_desired_generation.sql
@@ -1,0 +1,7 @@
+-- +goose Up
+ALTER TABLE hosts
+    ADD COLUMN desired_generation INT NOT NULL DEFAULT 0;
+
+-- +goose Down
+ALTER TABLE hosts
+    DROP COLUMN desired_generation;